winpthreads: correctly print time_t values in pthread_rwlock/rwlock{7,8}.c
These tests use %ld conversion specifier to print time_t values,
which may lead to incorrect result with 64-bit time_t since long
is always 32-bit on Windows.

In order to support old CRTs which may not support printing long long values,
convert time_t value to double and print it as a floating point number.

+++ mingw-w64-libraries/winpthreads/tests/pthread_rwlock/rwlock8.c
@@ -198,11 +198,11 @@ main (int argc, char *argv[])
 
   _ftime(&currSysTime2);
 
-  printf( "\nstart: %ld/%d, stop: %ld/%d, duration:%ld\n",
-          currSysTime1.time,currSysTime1.millitm,
-          currSysTime2.time,currSysTime2.millitm,
-          (currSysTime2.time*1000+currSysTime2.millitm) -
-          (currSysTime1.time*1000+currSysTime1.millitm));
+  printf( "\nstart: %.0f/%d, stop: %.0f/%d, duration:%.0f\n",
+          (double) currSysTime1.time,currSysTime1.millitm,
+          (double) currSysTime2.time,currSysTime2.millitm,
+          (double) (currSysTime2.time*1000+currSysTime2.millitm) -
+          (double) (currSysTime1.time*1000+currSysTime1.millitm));
 
   return 0;
 }
